+++ powerdoc/man.py
@@ -2,9 +2,11 @@ from collections.abc import Generator
 from datetime import date
 from typing import Any
 
-from attrs import define
+from attrs import define, field
 
 from powercli.command import Command
+from powercli.args import Flag
+from powercli._help_utils import _add_description
 
 from .builder import Builder
 
@@ -23,10 +25,32 @@ def escape(text: str) -> str:
     text = "".join(lines)
     return text
 
+def _prefix_flag(command: Command[Any, Any], flag: Flag[Any, Any, Any], *, visible_aliases: bool = True, hidden_aliases: bool = False) -> Generator[str, None, None]:
+    """Generates each name of a flag with the matching prefix."""
+    if flag.short is not None and command.prefix_short is not None:
+        yield command.prefix_short + flag.short
+    if flag.long is not None and command.prefix_long is not None:
+        yield command.prefix_long + flag.long
+    if visible_aliases:
+        for alias in flag.short_aliases:
+            if command.prefix_short is not None:
+                yield command.prefix_short + alias
+        for alias in flag.long_aliases:
+            if command.prefix_long is not None:
+                yield command.prefix_long + alias
+    if hidden_aliases:
+        for alias in flag.short_hidden_aliases:
+            if command.prefix_short is not None:
+                yield command.prefix_short + alias
+        for alias in flag.long_hidden_aliases:
+            if command.prefix_long is not None:
+                yield command.prefix_long + alias
+
 
 @define
 class ManBuilder(Builder):
     command: Command[Any, Any]
+    section: int = field(kw_only=True)
 
     def build(self) -> None:
         lines = []
@@ -36,12 +60,13 @@ class ManBuilder(Builder):
                 *self.make_name(),
                 *self.make_synopsis(),
                 *self.make_description(),
+                *self.make_options(),
             ]
         )
         print("\n".join(lines))
 
     def make_title_line(self) -> Generator[str, None, None]:
-        yield f'.TH "{escape(self.command.name.upper())}" "1" "{date.today():%b %Y}" "{escape(self.command.name)}"'
+        yield f'.TH "{escape(self.command.name.upper())}" "{self.section}" "{date.today():%b %Y}" "{escape(self.command.name)}"'
 
     def make_name(self) -> Generator[str, None, None]:
         yield ".SH NAME"
@@ -53,6 +78,7 @@ class ManBuilder(Builder):
         for flag in self.command._flags:
             yield "["
             names = []
+            # TODO: use _prefix_flag()
             if self.command.prefix_short is not None:
                 if (short_name := flag.short) is not None:
                     names.append(self.command.prefix_short + short_name)
@@ -70,3 +96,20 @@ class ManBuilder(Builder):
         if (desc := self.command.description) is not None:
             yield ".SH DESCRIPTION"
             yield escape(desc)
+
+    def make_options(self) -> Generator[str, None, None]:
+        # TODO: escape
+        indent = 7
+        if self.command.has_flag():
+            yield ".SH OPTIONS"
+            for flag in self.command._flags:
+                flag_names = ", ".join(_prefix_flag(self.command, flag))
+                text = _add_description(
+                    flag_names,
+                    indent=0,
+                    description_indentation=indent,
+                    description=flag.description,
+                    long_description=flag.long_description
+                )
+                yield text
+                yield ""